+++ packages/tpetra/core/ext/TpetraExt_MatrixMatrix_ExtraKernels_decl.hpp
@@ -98,25 +98,8 @@ static void jacobi_A_B_reuse_LowThreadGustavsonKernel(Scalar omega,
                                                       Teuchos::RCP<const Import<LocalOrdinal, GlobalOrdinal, Tpetra::KokkosCompat::KokkosOpenMPWrapperNode> > Cimport,
                                                       const std::string& label,
                                                       const Teuchos::RCP<Teuchos::ParameterList>& params);
-#endif
-
-/***************************** Matrix-Matrix Generic Kernels *****************************/
-template <class Scalar, class LocalOrdinal, class GlobalOrdinal, class Node, class LocalOrdinalViewType>
-static void jacobi_A_B_newmatrix_MultiplyScaleAddKernel(Scalar omega,
-                                                        const Vector<Scalar, LocalOrdinal, GlobalOrdinal, Node>& Dinv,
-                                                        CrsMatrixStruct<Scalar, LocalOrdinal, GlobalOrdinal, Node>& Aview,
-                                                        CrsMatrixStruct<Scalar, LocalOrdinal, GlobalOrdinal, Node>& Bview,
-                                                        const LocalOrdinalViewType& Acol2rrow,
-                                                        const LocalOrdinalViewType& Acol2Irow,
-                                                        const LocalOrdinalViewType& Bcol2Ccol,
-                                                        const LocalOrdinalViewType& Icol2Ccol,
-                                                        CrsMatrix<Scalar, LocalOrdinal, GlobalOrdinal, Node>& C,
-                                                        Teuchos::RCP<const Import<LocalOrdinal, GlobalOrdinal, Node> > Cimport,
-                                                        const std::string& label,
-                                                        const Teuchos::RCP<Teuchos::ParameterList>& params);
 
 /***************************** Triple Product OpenMP Only Kernels *****************************/
-#ifdef HAVE_TPETRA_INST_OPENMP
 template <class Scalar, class LocalOrdinal, class GlobalOrdinal, class LocalOrdinalViewType>
 static void mult_R_A_P_newmatrix_LowThreadGustavsonKernel(CrsMatrixStruct<Scalar, LocalOrdinal, GlobalOrdinal, Tpetra::KokkosCompat::KokkosOpenMPWrapperNode>& Rview,
                                                           CrsMatrixStruct<Scalar, LocalOrdinal, GlobalOrdinal, Tpetra::KokkosCompat::KokkosOpenMPWrapperNode>& Aview,
